+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
     "python.testing.pytestArgs": [
-        "Playwright-tests"
+        "Playwright"
     ],
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true
