--- Playwright-tests/pages/kevin/home_page.py
@@ -1,1 +1,0 @@
-kevin_home_page_paragraph_xpath = "xpath=/html/body/app-root/app-home/p"
