--- Playwright-tests/pages/pogoda/home_page.py
@@ -1,1 +1,0 @@
-pogoda_home_page_paragraph_xpath = "xpath=/html/body/app-root/app-home/p"
