+++ Playwright/pages/kevin/home_page.py
@@ -1,0 +1,1 @@
+home_page_paragraph_xpath = "xpath=/html/body/app-root/app-home/p"
