+++ Playwright/steps/test_kevin_home_page.py
@@ -1,9 +1,9 @@
 import pytest
 from pytest_bdd import scenarios, given, then, parsers
 from playwright.sync_api import sync_playwright, expect
-from ..pages.pogoda.home_page import pogoda_home_page_paragraph_xpath
+from Playwright.pages.kevin.home_page import home_page_paragraph_xpath
 
-scenarios("../features/pogoda_home_page.feature")
+scenarios("../features/kevin_home_page.feature")
 
 @pytest.fixture
 def page():
@@ -13,11 +13,11 @@ def page():
         yield page
         browser.close()
 
-@given("I navigate to pogoda home page")
-def navigate_to_pogoda_home_page(page):
-    page.goto("http://localhost:80/pogoda")
+@given("I navigate to kevin home page")
+def navigate_to_kevin_home_page(page):
+    page.goto("http://localhost:80/kevin")
 
 @then(parsers.parse('it should display the text "{text}"'))
 def verify_text(page, text):
-    locator = page.locator(pogoda_home_page_paragraph_xpath)
+    locator = page.locator(home_page_paragraph_xpath)
     expect(locator).to_have_text(text)
